make /trade index
