+++ components/Nav/Navbar/index.tsx
@@ -341,7 +341,7 @@ const NavBarContent: React.FC = styled(({ className }) => {
                     {/*<Link href="/trade/basic">*/}
                     {/*    <a className="m-auto">Trade</a>*/}
                     {/*</Link>*/}
-                    <Link href="/trade">
+                    <Link href="/">
                         <a className="m-auto">Trade</a>
                     </Link>
                 </li>
